Rename Main to Landing folder
For better understanding of project structure, the main folder has been renamed since main is used by many pages as a html semantic elementand not just for the landing page.
