+++ src/Components/Landing/AboutSection.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "../../styles/main/about.css";
+import "../../styles/landing/about.css";
 
 export default function AboutSection() {
   return (
