+++ src/Components/LandingPage.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Footer from "./Footer";
 import Header from "./Header";
-import Main from "./Main/Main";
+import Main from "./Landing/Main";
 
 export default function LandingPage() {
   return (
